+++ integration/test/app/basic/basic_test.go
@@ -9,10 +9,13 @@ import (
 
 	"github.com/giantswarm/apiextensions/v3/pkg/crd"
 	"github.com/giantswarm/apiextensions/v3/pkg/label"
+	"github.com/giantswarm/appcatalog"
 	"github.com/giantswarm/apptest"
 	"github.com/giantswarm/backoff"
 	"github.com/giantswarm/helmclient/v4/pkg/helmclient"
+	"github.com/spf13/afero"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/yaml"
 
 	"github.com/giantswarm/app-operator/v3/integration/key"
 	"github.com/giantswarm/app-operator/v3/integration/templates"
@@ -21,8 +24,8 @@ import (
 // TestAppLifecycle tests a chart CR can be created, updated and deleted
 // uaing a app, appCatalog CRs processed by app-operator.
 //
-// - Create appcatalog and app CRs.
 // - Install chart-operator.
+// - Create test app CR.
 // - Ensure chart CR specified in the app CR is deployed.
 //
 // - Update app version in app CR.
@@ -36,15 +39,59 @@ func TestAppLifecycle(t *testing.T) {
 	var err error
 
 	{
-		crdName := "Chart"
-		config.Logger.Debugf(ctx, "ensuring %#q CRD exists", crdName)
+		{
+			crdName := "Chart"
+			config.Logger.Debugf(ctx, "ensuring %#q CRD exists", crdName)
 
-		err := config.K8sClients.CRDClient().EnsureCreated(ctx, crd.LoadV1("application.giantswarm.io", crdName), backoff.NewMaxRetries(7, 1*time.Second))
+			err := config.K8sClients.CRDClient().EnsureCreated(ctx, crd.LoadV1("application.giantswarm.io", crdName), backoff.NewMaxRetries(7, 1*time.Second))
+			if err != nil {
+				t.Fatalf("expected %#v got %#v", nil, err)
+			}
+
+			config.Logger.Debugf(ctx, "ensured %#q CRD exists", crdName)
+		}
+
+		var tarballPath string
+		{
+			config.Logger.Debugf(ctx, "installing %#q", key.ChartOperatorUniqueName())
+
+			tarballURL, err := appcatalog.GetLatestChart(ctx, key.DefaultCatalogStorageURL(), key.ChartOperatorName(), key.ChartOperatorVersion())
+			if err != nil {
+				t.Fatalf("expected %#v got %#v", nil, err)
+			}
+
+			tarballPath, err = config.HelmClient.PullChartTarball(ctx, tarballURL)
+			if err != nil {
+				t.Fatalf("expected %#v got %#v", nil, err)
+			}
+
+			defer func() {
+				fs := afero.NewOsFs()
+				err := fs.Remove(tarballPath)
+				if err != nil {
+					t.Fatalf("expected %#v got %#v", nil, err)
+				}
+			}()
+		}
+
+		var values map[string]interface{}
+		{
+			err = yaml.Unmarshal([]byte(templates.ChartOperatorValues), &values)
+			if err != nil {
+				t.Fatalf("expected %#v got %#v", nil, err)
+			}
+		}
+
+		opts := helmclient.InstallOptions{
+			ReleaseName: key.ChartOperatorUniqueName(),
+			Wait:        true,
+		}
+		err = config.HelmClient.InstallReleaseFromTarball(ctx, tarballPath, key.Namespace(), values, opts)
 		if err != nil {
 			t.Fatalf("expected %#v got %#v", nil, err)
 		}
 
-		config.Logger.Debugf(ctx, "ensured %#q CRD exists", crdName)
+		config.Logger.Debugf(ctx, "installed %#q", key.ChartOperatorUniqueName())
 	}
 
 	{